+++ src/main/java/com/wootube/ioi/domain/repository/VideoRepository.java
@@ -1,6 +1,6 @@
 package com.wootube.ioi.domain.repository;
 
-import com.wootube.ioi.domain.model.video.Video;
+import com.wootube.ioi.domain.model.Video;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
